fix: suggested tags query speed (#280)
* feat: utility function for reordering data

* feat: tags dataloader + context initialization

* feat: improve performance of suggested tags query

In practice, only need the 'name' or 'id' values in
the response for suggestedTags on a SavedItem. Update
the logic so that other Tag fields are lazily resolved
only if they are requested by the client. The field
values are only fetched from the database if they are not
already provided by the parent in the resolver chain.
To avoid multiple database calls, use a dataloader to
batch load the Tag objects.

This avoids an additional join on the suggested tags
query, which restores performance to the level of the
v3 REST API (now using the same query).

* chore: use reorderResultByKey in all dataloader functions