feat(tags): enable tagging by givenUrl
Resolve the itemId from the url by requesting
from Parser. This is because givenUrl is not indexed
in the list table, so any url-based lookup must
scan an entire user's list. This service has access
to the parser database which stores urls and the
itemIds assigned to them, but the url-processing logic
is complex (we get around having to reimplement it
by simply invoking the parser). The long-term solution
is to add an index to givenUrl to make this access
pattern efficient. This request has been made to SRE
but is dependent on POC-194.

Helps mobile clients working offline, since they
do not have access to the server-generated itemId
used as the key for all other queries/mutations
on saved items.

[IN-1453]