fix(db): chunk queries to avoid full table scan
If 'where in' array is too large, then
mysql will ignore indices and scan the
entire table. The exact value at which
this happens depends on some memory settings.
Manual testing showed that 10k size will work
(we can tune this with the config value in
the future).

env var for batch size