fix(tags): display tags applied to many saves (#800)
* chore(tags): add failing test for tags not showing up

Case happens when the concatenated field
of saved item ids goes beyond a certain length;
then the result gets filtered out by the dataloader
which is looking for an ID that might be missing
due to truncated field. This happens for tags applied
to many items (and the item_ids are high cardinality)

[IN-1510]
"

* fix(tags): ensure all tags are returned

Fixes a big where tags do not appear
if applied to many items. Remove the
concatenated list of itemIds with the
response and resolve savedItems directly.

[IN-1510]

* chore: self-review fixes